Sync
Former-commit-id: 0830ce843df1168cca91abefddee3fe58706258d [formerly 926a4fac86e8f1ed84b0d3659577892f9f68daed]
Former-commit-id: fdb91e5e6b8e03f3ae9ef34f7fe9105d4c427391

+++ fancy_core/CommandContextDX12.h
@@ -63,7 +63,7 @@ namespace Fancy {
 
     void SetResourceTransitionBarrier(const GpuResource* aResource, D3D12_RESOURCE_STATES aNewState) const;
     void SetResourceTransitionBarriers(const GpuResource** someResources, const D3D12_RESOURCE_STATES* someNewStates, uint aNumStates) const;
-    void SetSubresourceTransitionBarrier(const GpuResource* aResource, uint aSubresourceIndex, D3D12_RESOURCE_STATES aNewState) const;
+    void SetSubresourceTransitionBarrier(const GpuResource* aResource, uint16 aSubresourceIndex, D3D12_RESOURCE_STATES aNewState) const;
     void SetSubresourceTransitionBarriers(const GpuResource** someResources, const D3D12_RESOURCE_STATES* someNewStates, const uint16** someSubresourceLists, const uint* someNumSubresources, uint aNumStates) const;
     void SetResourceUAVbarrier(const GpuResource* aResource) const;
